# aep-marketplace/public/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "MarketNow Skills API",
    "description": "The open marketplace for MCP-compatible agent skills. 8,535 verified skills, micro-priced from $0.99 to $9.99. Designed for autonomous agent consumption.\n\n## Authentication\nMost read endpoints are public (no auth required). Write endpoints (checkout) require no auth but redirect to Stripe.\n\n## Rate Limits\n- Anonymous: 60 requests/minute\n- The full catalog (/api/skills.json) is ~7MB — cache locally for 24h\n\n## Languages\nSkills are available in 5 languages: en, es, zh, pt, fr. Each skill with a system_prompt has a `translations` field.",
    "version": "3.0.0",
    "contact": {
      "name": "MarketNow",
      "url": "https://marketnow.site",
      "email": "eddyflores100@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://marketnow.site",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/mandates": {
      "get": {
        "summary": "Get / list delegated mandates",
        "description": "ACP / AP2 delegated mandates. With ?id= returns a single mandate. With ?owner=0x... returns all mandates owned by that wallet. With ?agent=agent_xxx returns all mandates granted to that agent.",
        "operationId": "getMandates",
        "parameters": [
          { "name": "id", "in": "query", "required": false, "schema": { "type": "string" }, "description": "mand_xxx" },
          { "name": "owner", "in": "query", "required": false, "schema": { "type": "string" }, "description": "wallet address (lowercase)" },
          { "name": "agent", "in": "query", "required": false, "schema": { "type": "string" }, "description": "agentId" }
        ],
        "responses": {
          "200": { "description": "Mandate(s) returned" },
          "404": { "description": "Not found" }
        }
      },
      "post": {
        "summary": "Create / revoke / spend a mandate",
        "description": "Three actions selected by the 'action' field in the body:\n- (no action): create a new mandate\n- action=revoke: revoke a mandate by id\n- action=spend: record a spend against a mandate (called internally by /api/agent-purchase after a successful on-chain payment)",
        "operationId": "postMandates",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": { "type": "string", "enum": ["revoke", "spend"] },
                  "owner": { "type": "string", "description": "wallet address (create only)" },
                  "agentId": { "type": "string", "description": "create only" },
                  "agentName": { "type": "string", "description": "create only, optional" },
                  "spendingLimitUsd": { "type": "number", "minimum": 0.01, "maximum": 500, "description": "create only" },
                  "perPurchaseCapUsd": { "type": "number", "minimum": 0.01, "maximum": 50, "description": "create only, optional" },
                  "categories": { "type": "array", "items": { "type": "string" }, "description": "create only, optional (defaults to ['*'])" },
                  "expiresAt": { "type": "string", "format": "date-time", "description": "create only, optional" },
                  "signature": { "type": "string", "description": "EIP-191 signature over mandate hash, optional but recommended" },
                  "id": { "type": "string", "description": "required for revoke / spend" },
                  "amount": { "type": "number", "description": "required for spend" },
                  "txHash": { "type": "string", "description": "optional for spend, on-chain tx reference" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Action completed" },
          "201": { "description": "Mandate created" },
          "400": { "description": "Validation error" },
          "404": { "description": "Mandate not found" },
          "409": { "description": "Mandate revoked / expired / insufficient remaining" }
        }
      }
    },
...
  },
...
}